refactor: interaction_commands.py 4개 클래스 → 개별 파일 분리 (태스크 29.3)
- emote_command.py, give_command.py, follow_command.py, players_command.py
- interaction_commands.py 하위 호환성 re-export 유지
